# openapi.yaml
openapi: 3.1.0
info:
  title: Домовой — API приложения дома в MAX
  version: 1.0.0
  description: 'API мини-приложения и бота MAX «Домовой»: вход по QR-коду квитанции, обращения в УК, начисления,
    счётчики, лента и опросы дома, раздел председателя, кабинет УК, подбор дома.


    **Как войти для проверки.** Без квитанции и без MAX — ролью демо-дома: `POST /api/demo/roles/{key}/take` (key:
    chairman, owner12, member12, owner45, owner78, newcomer30) вернёт `token`; дальше заголовок `Authorization:
    Bearer <token>`. Кабинет УК — `POST /api/dispatcher/login` с логином демо-УК (пароль — на служебном слайде презентации),
    тоже `token`.


    **Ошибки** — JSON `{ "error": "код", "message": "текст для человека" }`.


    **Модель доступа.** Квитанция — заявка, а не пропуск: новая квартира приходит со статусом `pending`. Уровень
    0 (pending) — свои начисления, счётчики и обращения в УК сразу. Уровень 1 (active) — соседи, лента, опросы.
    Уровень 2 — дом целиком, только председателю. Обращение удалить нельзя.


    Кабинет оператора (`/api/admin/*`) — внутренний и здесь не описан.'
servers:
- url: https://domovoy.devcore.com.ru
  description: Бой
- url: http://localhost:3000
  description: Локальный запуск (docker compose)
...
security:
- bearerAuth: []
- cookieAuth: []
paths:
...
  /api/bills/{billId}/paid:
    post:
      tags:
      - Житель
      summary: Отметить начисление оплаченным
      description: 'Отметка жителя: приложение не видит платёж, это его дневник.'
      operationId: post_api_bills_billId_paid
      parameters:
      - name: billId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    const: ok
                  billStatus:
                    type: string
                required:
                - status
                - billStatus
        '400':
          description: Неверные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Нет прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                paid:
                  type: boolean
                kopecks:
                  type: integer
  /api/bills/{billId}/pay-qr:
    post:
      tags:
      - Житель
      summary: Ссылка на QR для оплаты
      description: Подписанная ссылка на картинку QR квитанции на 15 минут — только тому, кто принёс квитанцию.
      operationId: post_api_bills_billId_pay_qr
      parameters:
      - name: billId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: {}
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type:
                    - string
                    - 'null'
                  fileName:
                    type:
                    - string
                    - 'null'
                required:
                - url
                - fileName
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Нет прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/bills/{billId}/qr.png:
    get:
      tags:
      - Житель
      summary: Картинка QR квитанции
      description: По подписанной ссылке из pay-qr.
      operationId: get_api_bills_billId_qr_png
      parameters:
      - name: billId
        in: path
        required: true
        schema:
          type: string
      - name: t
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Успех
          content:
            image/png:
              schema:
                type: string
                format: binary
        '403':
          description: Нет прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security: []
...
  /api/meters/{meterId}/readings:
    post:
      tags:
      - Житель
      summary: Записать показание
      description: Проверяет, что показание не меньше прошлого и не выбивается из обычного расхода.
      operationId: post_api_meters_meterId_readings
      parameters:
      - name: meterId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                value:
                  type:
                  - string
                  - 'null'
              required:
              - value
      responses:
        '201':
          description: Создано
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  consumption:
                    type:
                    - number
                    - 'null'
                  first:
                    type:
                    - boolean
                    - 'null'
                  warnings:
                    type: array
                    items:
                      type: object
                      properties:
                        code:
                          type:
                          - string
                          - 'null'
                        message:
                          type:
                          - string
                          - 'null'
                      required:
                      - code
                      - message
                required:
                - status
                - consumption
                - first
                - warnings
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/notifications:
    get:
      tags:
      - Уведомления
      summary: Уведомления
      description: Смена статуса обращения, вопросы УК, отключения.
      operationId: get_api_notifications
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  unread:
                    type:
                    - number
                    - 'null'
                  total:
                    type: number
                  notifications:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        kind:
                          type:
                          - string
                          - 'null'
                        title:
                          type:
                          - string
                          - 'null'
                        body:
                          type:
                          - string
                          - 'null'
                        read:
                          type:
                          - boolean
                          - 'null'
                        at:
                          type:
                          - string
                          - 'null'
                          format: date-time
                        delivered:
                          type:
                          - boolean
                          - 'null'
                        deepLinkPayload:
                          type:
                          - string
                          - 'null'
                      required:
                      - id
                      - kind
                      - title
                      - body
                      - read
                      - at
                      - delivered
                      - deepLinkPayload
                required:
                - unread
                - total
                - notifications
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/notifications/read:
    post:
      tags:
      - Уведомления
      summary: Прочитать уведомления
      operationId: post_api_notifications_read
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                required:
                - status
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/notifications/settings:
    get:
      tags:
      - Уведомления
      summary: Настройки уведомлений
      description: Что присылать сообщением бота MAX.
      operationId: get_api_notifications_settings
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  mode:
                    type:
                    - string
                    - 'null'
                  kinds:
                    type: object
                    properties:
                      request_status:
                        type:
                        - boolean
                        - 'null'
                      request_assigned:
                        type:
                        - boolean
                        - 'null'
                      access_request:
                        type:
                        - boolean
                        - 'null'
                      outage:
                        type:
                        - boolean
                        - 'null'
                      meters_reminder:
                        type:
                        - boolean
                        - 'null'
                      payment_reminder:
                        type:
                        - boolean
                        - 'null'
                      review_invite:
                        type:
                        - boolean
                        - 'null'
                    required:
                    - request_status
                    - request_assigned
                    - access_request
                    - outage
                    - meters_reminder
                    - payment_reminder
                    - review_invite
                  available:
                    type: array
                    items:
                      type: object
                      properties:
                        kind:
                          type:
                          - string
                          - 'null'
                        label:
                          type:
                          - string
                          - 'null'
                        hint:
                          type:
                          - string
                          - 'null'
                        important:
                          type:
                          - boolean
                          - 'null'
                      required:
                      - kind
                      - label
                      - hint
                      - important
                required:
                - mode
                - kinds
                - available
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
      - Уведомления
      summary: Изменить настройки уведомлений
      operationId: post_api_notifications_settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mode:
                  type:
                  - string
                  - 'null'
                kinds:
                  type: object
                  properties:
                    payment_reminder:
                      type:
                      - boolean
                      - 'null'
                    request_status:
                      type:
                      - boolean
                      - 'null'
                    request_assigned:
                      type:
                      - boolean
                      - 'null'
                    access_request:
                      type:
                      - boolean
                      - 'null'
                    outage:
                      type:
                      - boolean
                      - 'null'
                    meters_reminder:
                      type:
                      - boolean
                      - 'null'
                  required:
                  - payment_reminder
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  mode:
                    type:
                    - string
                    - 'null'
                  kinds:
                    type: object
                    properties:
                      request_status:
                        type:
                        - boolean
                        - 'null'
                      request_assigned:
                        type:
                        - boolean
                        - 'null'
                      access_request:
                        type:
                        - boolean
                        - 'null'
                      outage:
                        type:
                        - boolean
                        - 'null'
                      meters_reminder:
                        type:
                        - boolean
                        - 'null'
                      payment_reminder:
                        type:
                        - boolean
                        - 'null'
                      review_invite:
                        type:
                        - boolean
                        - 'null'
                    required:
                    - request_status
                    - request_assigned
                    - access_request
                    - outage
                    - meters_reminder
                    - payment_reminder
                    - review_invite
                required:
                - status
                - mode
                - kinds
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/properties/{propertyId}/analytics:
    get:
      tags:
      - Житель
      summary: Аналитика расходов
      description: Суммы по месяцам и услугам, сравнение с прошлым периодом.
      operationId: get_api_properties_propertyId_analytics
      parameters:
      - name: propertyId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  payments:
                    type: object
                    properties:
                      points:
                        type: array
                        items:
                          type: object
                          properties:
                            period:
                              type:
                              - string
                              - 'null'
                            label:
                              type:
                              - string
                              - 'null'
                            value:
                              type:
                              - number
                              - 'null'
                            formatted:
                              type:
                              - string
                              - 'null'
                          required:
                          - period
                          - label
                          - value
                          - formatted
                      change: {}
                      months:
                        type:
                        - number
                        - 'null'
                      latest:
                        type: object
                        properties:
                          period:
                            type:
                            - string
                            - 'null'
                          parts:
                            type: array
                            items:
                              type: object
                              properties:
                                label:
                                  type:
                                  - string
                                  - 'null'
                                persAcc:
                                  type:
                                  - string
                                  - 'null'
                                kopecks:
                                  type:
                                  - number
                                  - 'null'
                                formatted:
                                  type:
                                  - string
                                  - 'null'
                              required:
                              - label
                              - persAcc
                              - kopecks
                              - formatted
                          accountsTotal:
                            type:
                            - number
                            - 'null'
                          partial:
                            type:
                            - boolean
                            - 'null'
                        required:
                        - period
                        - parts
                        - accountsTotal
                        - partial
                    required:
                    - points
                    - change
                    - months
                    - latest
                  meters:
                    type: array
                    items: {}
                  forecast:
                    type: object
                    properties:
                      kopecks:
                        type:
                        - number
                        - 'null'
                      formatted:
                        type:
                        - string
                        - 'null'
                      basis:
                        type:
                        - string
                        - 'null'
                    required:
                    - kopecks
                    - formatted
                    - basis
                  hint:
                    type:
                    - string
                    - 'null'
                  tips:
                    type: array
                    items:
                      type: object
                      properties:
                        title:
                          type:
                          - string
                          - 'null'
                        body:
                          type:
                          - string
                          - 'null'
                      required:
                      - title
                      - body
                required:
                - payments
                - meters
                - forecast
                - hint
                - tips
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Токен сессии из /api/auth/qr, /api/auth/max, /api/demo/roles/{key}/take или /api/dispatcher/login
    cookieAuth:
      type: apiKey
      in: cookie
      name: zd_session
      description: Та же сессия в cookie — её ставит сервер при входе
...
  schemas:
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: string
          description: Машинный код ошибки
        message:
          type: string
          description: Понятный человеку текст
  responses:
    Unauthorized:
      description: Нужен вход
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'